Add minimal error handler for Net::Http2 client reqeust
This is to prevent things like Errno::ECONNRESET from terminating the
dispatcher in daemon mode, which leaves the notifications stuck
forever in the processing state.

Perhaps more can be done here, whether to mark the notification for
retry or failure is perhaps something for the user to configure.

See https://github.com/ostinelli/net-http2/issues/23#issuecomment-361616710